Move main classes to Models folder
