+++ Controllers/ListingController.cs
@@ -6,6 +6,7 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using Karma.Models;
 
 namespace Karma.Controllers
 {
