+++ Models/Listing.cs
@@ -1,7 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Karma	{ 
+namespace Karma.Models
+{ 
 	public class Listing
 	{
 		public string Id { get; set; }
