+++ Models/User.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 
 
-namespace Karma
+namespace Karma.Models
 {
 	public class User
 	{
